Update to current version

+++ PALMSDPU source code/src/edu/ucsd/cwphs/palms/poi/DAOpoi.java
@@ -9,8 +9,23 @@ import java.util.Date;
 
 import edu.ucsd.cwphs.palms.util.EventLogger;
 
+/**
+ * DAOpoi	-	Data Access Object to Point of Information (POI) database
+ * 
+ * @author Fredric Raab, UCSD CWPHS		fraab@ucsd.edu
+ * @version 1.2.0
+ * @since 2015-05-28
+ * 
+ * History:
+ * 	2015.03.31		1.0.0		Initial Release
+ * 	2015.05.27		1.1.0		Fails-safe when initial connection to database fails
+ * 								in connect(), add code to load postgresql driver
+ *  2016.04.06		1.2.0		Remove localCache, add buffer and md2k_types to database
+ */
+
 public class DAOpoi {
 
+	// replace with parameters for your database
 	private final String databaseURL = "jdbc:postgresql://137.110.115.96/md2kPOI";
 	private final String databaseUser = "md2k";
 	private final String databasePassword = "md2k";
@@ -19,28 +34,19 @@ public class DAOpoi {
 	private ResultSet rs = null;
 	private PreparedStatement ps = null;
 	
-	private static ArrayList<POI> localCache = new ArrayList<POI>();
-	private static int cacheHits = 0;
+	// stats
 	private static int inserts = 0;
+	
+	// TODO: need to return error message on connection failure
 	
 	public DAOpoi(){
 		connect();
 	}
 	
 	public ArrayList<POI> getNearBy(Double lat, Double lon, int radius, String types){
-		ArrayList<POI> results = new ArrayList<POI>();
-		
-		for (int i=0; i<localCache.size(); i++){
-			POI poi= localCache.get(i);
-				if (poi.isType(types))
-					if (poi.isNearBy(lat, lon, radius)){
-					results.add(poi);
-					cacheHits++;
-					}
-		}
+
 		// query DB
-		results = findPOI(lat, lon, radius, types);
-		
+		ArrayList<POI> results = findPOI(lat, lon, radius, types);
 		
 		return results;
 	}
@@ -48,7 +54,6 @@ public class DAOpoi {
 	public boolean insertPOIs(ArrayList<POI> poiList){
 		for (int i=0; i<poiList.size(); i++){
 			POI poi = poiList.get(i);
-			localCache.add(poi);
 			insertPOI(poi);
 			inserts++;
 		}
@@ -57,7 +62,7 @@ public class DAOpoi {
 	
 	public int dbSize(){
 		int count = 0;
-		count = localCache.size();
+		// TODO:  Implemement function
 		return count;
 	}
 	
@@ -73,20 +78,17 @@ public class DAOpoi {
 		return true;
 	}
 	
-	public int getCacheHitCount(){
-		return cacheHits;
-	}
-	
 	public int getInsertsCount(){
 		return inserts;
 	}
 	
 	public boolean connect(){
 		try {
+			Class.forName("org.postgresql.Driver");
 			con = DriverManager.getConnection(databaseURL, databaseUser, databasePassword);
 		}
 		catch (Exception ex){
-			System.out.println("DAOpoi Exception:" + ex.getMessage());
+			EventLogger.logException("DAOpoi.connect - connection to database failed: ", ex);
 			con = null;
 			return false;
 		}
@@ -95,9 +97,12 @@ public class DAOpoi {
 	
 	public boolean insertPOI(POI poi){
 		boolean rc = true;
+		if (con == null)
+			return false;		// no connection to database
+		
 		try {
-			ps = con.prepareStatement("INSERT INTO poi (place_id, name, scope, vicinity, postal_address, date_created, date_expires, lon, lat, types, lonlatgeometry)" +
-									" VALUES (?,?,?,?,?,?,?,?,?,?, ST_GeometryFromText('POINT(" + poi.getLon() + " " + 
+			ps = con.prepareStatement("INSERT INTO poi (place_id, name, scope, vicinity, postal_address, date_created, date_expires, lon, lat, types, md2k_types, buffer, lonlatgeometry)" +
+									" VALUES (?,?,?,?,?,?,?,?,?,?,?,?, ST_GeometryFromText('POINT(" + poi.getLon() + " " + 
 									 poi.getLat() + ")', 4326));");
 			ps.setString(1, poi.getPlaceId());
 			ps.setString(2, poi.getName());
@@ -109,11 +114,14 @@ public class DAOpoi {
 			ps.setDouble(8, poi.getLon());			// lon first
 			ps.setDouble(9, poi.getLat());
 			ps.setString(10, poi.getTypes());
+			ps.setString(11, poi.getMd2kTypes());
+			ps.setInt(12, poi.getBuffer());
 			ps.execute();
 			EventLogger.logEvent("DAOpoi.insertPOI - inserted:" + poi.toCSV());
 		}
 		catch (Exception ex){
-			if (ex.getMessage().contains("duplicate key value"))
+			String message = ex.getMessage();		
+			if ((message != null) && message.contains("duplicate key value"))
 				rc = true;		// location already exists - return success
 			else {
 				EventLogger.logException("DAOpoi.insertPOI Exception:", ex);
@@ -128,9 +136,10 @@ public class DAOpoi {
 		ArrayList<POI> results = new ArrayList<POI>();
 		Double buffer = radius /1.0;					// turns int to double
 		
+		if (con == null)
+			return results;			// no connection to database
+		
 		try {
-//			ps = con.prepareStatement("Select * FROM poi WHERE ST_DWithin(lonlatgeometry, 'POINT(" + lon + " " + lat + ")',?);");
-//			ps = con.prepareStatement("Select * FROM poi WHERE ST_DWithin(lonlatgeometry, ST_GeomFromText('POINT(" + lon + " " + lat + ")', 4326),?);");
 			String sql = "Select * FROM poi WHERE ST_Distance_Sphere(lonlatgeometry, ST_GeomFromText('POINT(" + lon + " " + lat + ")', 4326)) <= ?";
 			sql = sql + parseTypes(types);
 			ps = con.prepareStatement(sql);
@@ -141,11 +150,17 @@ public class DAOpoi {
 				String name = rs.getString("name");
 				String scope = rs.getString("scope");
 				types = rs.getString("types");
+				String md2kTypes = rs.getString("md2k_types");
 				String vicinity = rs.getString("vicinity");
 				String postalAddress = rs.getString("postal_address");
-				lon = rs.getDouble("lon");
-				lat = rs.getDouble("lat");
-				POI poi = new POI(placeId, name, scope, types, vicinity, postalAddress, lat, lon);
+				double POIlon = rs.getDouble("lon");
+				double POIlat = rs.getDouble("lat");
+				int POIbuffer = rs.getInt("buffer");
+				POI poi = new POI(placeId, name, scope, types, md2kTypes, vicinity, postalAddress, POIlat, POIlon);
+				int distance = poi.getDistanceFrom(lat, lon);
+				poi.setDistance(distance);
+				poi.setBuffer(POIbuffer);
+				
 				results.add(poi);	
 			}
 		}
@@ -156,8 +171,51 @@ public class DAOpoi {
 		return results;
 	}
 	
+	
+	public ArrayList<POI> findPOIAlongRoute(double lat1, double lon1, double lat2, double lon2, int radius, String types){
+		ArrayList<POI> results = new ArrayList<POI>();
+		Double buffer = radius /1.0;					// turns int to double
+		
+		if (con == null)
+			return results;			// no connection to database
+		
+		try {
+			String sql = "Select * FROM poi WHERE (ST_Distance_Sphere(lonlatgeometry, ST_GeomFromText('LINESTRING(" + lon1 + " " + lat1 + 
+					", " + lon2 + " " + lat2 + ")', 4326)) <= ?) AND (name != 'No POI nearby') ";
+			sql = sql + parseTypes(types);
+			ps = con.prepareStatement(sql);
+			ps.setDouble(1, buffer);
+			rs = ps.executeQuery();
+			while (rs.next()){
+				String placeId = rs.getString("place_id");
+				String name = rs.getString("name");
+				String scope = rs.getString("scope");
+				types = rs.getString("types");
+				String md2kTypes = rs.getString("md2k_types");
+				String vicinity = rs.getString("vicinity");
+				String postalAddress = rs.getString("postal_address");
+				double POIlon = rs.getDouble("lon");
+				double POIlat = rs.getDouble("lat");
+				int POIbuffer = rs.getInt("buffer");
+				POI poi = new POI(placeId, name, scope, types, md2kTypes, vicinity, postalAddress, POIlat, POIlon);
+				int distance = poi.getDistanceFromRoute(lat1, lon1, lat2, lon2);
+				poi.setDistance(distance);
+				poi.setBuffer(POIbuffer);
+				
+				results.add(poi);	
+			}
+		}
+		catch (Exception ex){
+			System.out.println("DAOpoi.findPOIAlongRoute Exception:" + ex.getMessage());
+		}
+		closeAll();
+		return results;
+	}
+	
 	public ArrayList<POI> getAllPOIs(String types){
 		ArrayList<POI> results = new ArrayList<POI>();
+		if (con == null)
+			return results;				// no connection to database
 		
 		try {
 			String sql = "Select * FROM poi WHERE name NOT LIKE '"+ POI.NOPOINEARBY + "'";
@@ -169,11 +227,14 @@ public class DAOpoi {
 				String name = rs.getString("name");
 				String scope = rs.getString("scope");
 				types = rs.getString("types");
+				String md2kTypes = rs.getString("md2k_types");
 				String vicinity = rs.getString("vicinity");
 				String postalAddress = rs.getString("postal_address");
 				Double lon = rs.getDouble("lon");
 				Double lat = rs.getDouble("lat");
-				POI poi = new POI(placeId, name, scope, types, vicinity, postalAddress, lat, lon);
+				int buffer = rs.getInt("buffer");
+				POI poi = new POI(placeId, name, scope, types, md2kTypes, vicinity, postalAddress, lat, lon);
+				poi.setBuffer(buffer);
 				results.add(poi);	
 			}
 		}
@@ -187,6 +248,8 @@ public class DAOpoi {
 	
 	public String getGoogleType(String yahooType){
 		String result = null;
+		if (con == null)
+			return result;				// no connection to database
 		
 		try {
 			String sql = "Select googletype FROM typemapping WHERE type LIKE '?'";
@@ -206,6 +269,8 @@ public class DAOpoi {
 	
 	public boolean deleteExpired(Date date){
 		boolean rc = true;
+		if (con == null)
+			return false;				// no connection to database
 		try {
 			ps = con.prepareStatement("DELETE FROM poi WHERE date_expires < ?;");
 			ps.setDate(1, toSqlDate(date)); 
@@ -221,6 +286,8 @@ public class DAOpoi {
 	
 	public int countExpired(Date date){
 		int count = 0;
+		if (con == null)
+			return -1;				// no connection to database
 		try {
 			ps = con.prepareStatement("Select Count(*) FROM poi WHERE date_expires < ?;");
 			ps.setDate(1, toSqlDate(date)); 
@@ -238,6 +305,9 @@ public class DAOpoi {
 	
 	public boolean deleteScope(String scope){
 		boolean rc = true;
+		if (con == null)
+			return false;				// no connection to database
+		
 		try {
 			ps = con.prepareStatement("DELETE FROM poi WHERE scope = ?;");
 			ps.setString(1, scope); 
@@ -253,6 +323,8 @@ public class DAOpoi {
 	
 	public int countScope(String scope){
 		int count = 0;
+		if (con == null)
+			return -1;				// no connection to database
 		try {
 			if (scope == null) 
 				ps = con.prepareStatement("Select Count(*) FROM poi;");
